Cambia de nombre a componente Estadisticas y arregla css
